Revert "Move 1-bit backend under quantization"
This reverts commit 73d6218721f4cf14bbfc3706adca4bbca5a07991.
