--- mlx_vlm/quantization/__init__.py
@@ -1,1 +1,0 @@
-"""Weight quantization backends."""
