+++ mlx_vlm/utils.py
@@ -21,7 +21,7 @@ from transformers import AutoProcessor
 from transformers.processing_utils import ProcessorMixin
 
 from .models.base import BaseImageProcessor
-from .quantization.one_bit import _quantization_for_path, replace_one_bit_modules
+from .one_bit import _quantization_for_path, replace_one_bit_modules
 from .tokenizer_utils import load_tokenizer
 from .trainer.utils import apply_lora_layers
